% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Libraries.R
\name{check.valid.metadata.DwC}
\alias{check.valid.metadata.DwC}
\title{validate a DarwinCore class object}
\usage{
check.valid.metadata.DwC(d)
}
\arguments{
\item{d}{a DwC.event or DwC.occurrence object}
}
\value{
a boolean, TRUE when valid, FALSE when not
}
\description{
validates if a DwC.event or DwC.occurrence object is implemented correctly
}
\author{
Maxime Sweetlove CC-0 2020
}
