% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Libraries.R
\name{check.valid.MIxS.metadata}
\alias{check.valid.MIxS.metadata}
\title{validate a MIxS.metadata class object}
\usage{
check.valid.MIxS.metadata(d)
}
\arguments{
\item{d}{a MIxS.metadata object}
}
\value{
a boolean, TRUE when valid, FALSE when not
}
\description{
validates if a MIxS.metadata object is implemented correctly
}
\author{
Maxime Sweetlove CC-0 2020
}
