% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFormat_Utils.R
\name{combine.data}
\alias{combine.data}
\title{merge MIxS.metadata objects}
\usage{
combine.data(d1, d2, fill=NA, variables.as.cols=TRUE)
}
\arguments{
\item{d1}{a MIxS.metadata object}

\item{d2}{a MIxS.metadata object}

\item{fill}{character or NA. A value to put into the cells that have no data.}

\item{variables.as.cols}{boolean. If TRUE, the input data is assumed to have rows as samples and variables/parameters as columns. If FALSE the data is formatted the other was around. default TRUE}
}
\value{
a MIxS.metadata object
}
\description{
combine.data combines two MIxS.metadata objects into one, merging all common variables (columns as default).
}
\details{
Variables with matching names are merged, if variables.as.cols=TRUE this means columns are merged, if FALSE rows are merged. Any missing data as a result of the non-matchng variable-name will be filled by the fill argument.
}
\examples{
\dontrun{
combine.data(d1=test_MIxS, d2=test_MIxS2, fill=NA, variables.as.cols=TRUE)
}
}
\seealso{
Other formating functions: 
\code{\link{combine.data.frame}()},
\code{\link{eMoF.to.wideTable}()},
\code{\link{wideTab.to.hierarchicalTab}()},
\code{\link{wideTable.to.eMoF}()}
}
\author{
Maxime Sweetlove
}
\concept{formating functions}
