% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Utils.R
\name{commonTax.to.NCBI.TaxID}
\alias{commonTax.to.NCBI.TaxID}
\title{Convert a taxon name to an NCBI taxon ID}
\usage{
commonTax.to.NCBI.TaxID(taxon, fill.unknown = "12908")
}
\arguments{
\item{taxon}{character or character vector. The taxon names to be converted to NCBI tax IDs.}

\item{fill.unknown}{character. The string to return when a taxon was not found in the list with common taxa. Default is NCBI:txid12908 (the ID for "unknown sequences"), other options include NA or ""}
}
\value{
a character or character vector with the matching taxon IDs, with fill.unknown for those not found.
}
\description{
converts taxon names of common taxa (superkingdom and phylum level) to it's NCBI taxID using an internal library. For taxa not in the internal library, please see https://www.ncbi.nlm.nih.gov/Taxonomy/TaxIdentifier/tax_identifier.cgi
}
\details{
This function makes use of a limited internal library, not all the taxIDs are present
}
\seealso{
Other data archiving functions: 
\code{\link{FileNames.to.Table}()},
\code{\link{get.ENAName}()},
\code{\link{prep.metadata.ENA}()},
\code{\link{renameSequenceFiles}()},
\code{\link{sync.metadata.sequenceFiles}()}
}
\author{
Maxime Sweetlove
}
\concept{data archiving functions}
