% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Utils.R
\name{coordinate.to.decimal}
\alias{coordinate.to.decimal}
\title{convert a coordinate in degrees to decimal}
\usage{
coordinate.to.decimal(val)
}
\arguments{
\item{val}{a character string. A single latitude or longitude value to be transformed. Can include non-numeric character like the degree symbol, N-S-E-W wind directions,...}
}
\value{
a numeric value
}
\description{
Turns a latutude or longitude value in a degrees-minutes-seconds (DMS) format into a decimal value
}
\details{
N-S-W-E wind directions as well as degrees, minutes and seconds are recognized and turned into a numeric decimal coordinate value
}
\examples{
\donttest{
coordinate.to.decimal("40.33S")
coordinate.to.decimal("40<U+00B0>33S")
}
}
\seealso{
Other standardization functions: 
\code{\link{dataQC.DwC_general}()},
\code{\link{dataQC.DwC}()},
\code{\link{dataQC.MIxS}()}
}
\author{
Maxime Sweetlove
}
\concept{standardization functions}
