% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQC_Utils.R
\name{dataQC.completeTaxaNamesFromRegistery}
\alias{dataQC.completeTaxaNamesFromRegistery}
\title{Complete a list of taxonomic names with information form an authorotive taxonomic backbone}
\usage{
dataQC.completeTaxaNamesFromRegistery(taxaNames, taxBackbone="worms")
}
\arguments{
\item{taxaNames}{a character vector. A list with the taxonomic names to look for}

\item{taxBackbone}{a character string. The taxonomic backbone to querry. At present only "worms" is allowed.}
}
\value{
a dataframe with the following fields:scientificName, scientificNameID, aphID, kingdom, phylum, class, order, family, genus, specificEpithet, scientificNameAuthorship, namePublishedInYear
}
\description{
complete a list of taxonomic names by looking-up missing information on an accepted taxonomic registery
}
\details{
using the API-client connection to the World Registry of Marine Species (WORMS), additional taxonomic information can be added to an existing list of taxa. This function is a wrapper of the wm_name2id and wm_record functions in the worrms package
}
\seealso{
Other quality control functions: 
\code{\link{dataQC.LatitudeLongitudeCheck}()},
\code{\link{dataQC.TaxonListFromData}()},
\code{\link{dataQC.TermsCheck}()},
\code{\link{dataQC.dateCheck}()},
\code{\link{dataQC.eventStructure}()},
\code{\link{dataQC.findNames}()},
\code{\link{dataQC.generate.footprintWKT}()},
\code{\link{dataQC.guess.env_package.from.data}()},
\code{\link{dataQC.taxaNames}()}
}
\author{
Maxime Sweetlove CC-0 2020
}
\concept{quality control functions}
