% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQC_Main_DwC.R
\name{dataQC.DwC}
\alias{dataQC.DwC}
\title{format dataframes into a DarwinCore object}
\usage{
dataQC.DwC(Event=NA, Occurrence=NA, eMoF=NA, EML.url=NA, out.type="event", ask.input=TRUE)
}
\arguments{
\item{Event}{data.frame. A dataframe with the event data, structured with DarwinCore terms}

\item{Occurrence}{data.frame. A dataframe with the occurrence data, structured with DarwinCore terms}

\item{eMoF}{data.frame. A dataframe with the extended MeasurementOrFact (eMoF) data, structured with DarwinCore terms}

\item{EML.url}{character. A URL to the Ecological Metadata Language (EML) file.}

\item{out.type}{character. The type of DarwinCore of the output, either event or occurrence. If event, the output will have an event core with possible occurrence and eMoF extensions. If occurrence, the output will have an occurrence core with possibly an eMoF extension. Default event, if the parameter Event is not NA, out.type will be fixed as event.}

\item{ask.input}{logical. If TRUE, console input will be requested to the user when a problem occurs. Default TRUE}
}
\value{
a DwC.event (event core with occurrence and/or eMoF extensions) or DwC.occurence (occurrence core with eMoF extension) object
}
\description{
Takes one or more dataframes or a DarwinCore Archive and performs a basic Quality Controll. (see details)
}
\details{
DarwinCore is the biodiversity data standard develloped by TDWG, and is used by the Global Biodiversity Information Facility (GBIF). This function performs a basic and user-supervised quality control. This includes cheking all variables terms adhere to the DarwinCore vocabulary, listing other variables in an eMoF file or in the dynamicProperties field, and checking for obvious errors in the content of the data (typos, different NA values,...). All dataframes are combined into a DwC.event or DwC.occurrence object
}
\examples{
\dontrun{
dataQC.DwC(Event=event_data.frame, out.type="event")
dataQC.DwC(Occurrence=occurrence_data.frame, out.type="occurrence")
dataQC.DwC(Event=event_data.frame, eMoF=emof_data.frame, out.type="event")
dataQC.DwC(Event=event_data.frame, occurrence=occurrence_data.frame, out.type="event")
}
}
\seealso{
Other standardization functions: 
\code{\link{coordinate.to.decimal}()},
\code{\link{dataQC.DwC_general}()},
\code{\link{dataQC.MIxS}()}
}
\author{
Maxime Sweetlove CC-0 2020
}
\concept{standardization functions}
