% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQC_Main_DwCgeneral.R
\name{dataQC.DwC_general}
\alias{dataQC.DwC_general}
\title{Quality Controll (QC) for DarwinCore data}
\usage{
dataQC.DwC_general(dataset = NA, DwC.type = "event", ask.input = TRUE, complete.data=TRUE)
}
\arguments{
\item{dataset}{data.frame. A dataframe with the data, structured with DarwinCore terms}

\item{DwC.type}{character. The type of DarwinCore of the output, either event or occurrence. If event, the output will have an event core with possible occurrence and eMoF extensions. If occurrence, the output will have an occurrence core with possibly an eMoF extension. Default event, if the parameter Event is not NA, out.type will be fixed as event.}

\item{ask.input}{logical. If TRUE, console input will be requested to the user when a problem occurs. Default TRUE}

\item{complete.data}{logical. If TRUE, datathat has not been provided, but can be completed automatically will be added to the DarwinCore data.frame. For instance, footprintWKT can be generated from the coordinates, or higher taxonomic level names (like kingdom, phylum,...) can looked up with the species name. Defaut TRUE.}
}
\value{
a dataframe that is formatted as either: an event or occurrence
}
\description{
performs a general QC on a darwinCore file. Desired output is either and event (core) file, occurrence (core or extension) or extended Measurement or Fact (eMoF; extension) file. Not an Ecological dataset Language (EML) file.
}
\details{
DarwinCore is the biodiversity data standard develloped by TDWG, and is used by the Global Biodiversity Information Facility (GBIF). This function performs a basic and user-supervised quality control. This includes cheking all variables terms adhere to the DarwinCore vocabulary, listing other variables in an eMoF file or in the dynamicProperties field, and checking for obvious errors in the content of the data (typos, different NA values,...)
}
\examples{
\dontrun{
test_event <- data.frame(eventID=c("sample1", "sample2"),
                         eventDate=c("2021-09-27", "2021-09-28"),
                         decimalLatitude=c("54.7", "33"),
                         decimalLongitude=c("88.9", "-48.4"),
                         row.names=c("sample1", "sample2"))
dataQC.DwC_general(dataset=test_event, DwC.type = "event", complete.data=TRUE)
}
}
\seealso{
Other standardization functions: 
\code{\link{coordinate.to.decimal}()},
\code{\link{dataQC.DwC}()},
\code{\link{dataQC.MIxS}()}
}
\author{
Maxime Sweetlove CC-0 2020
}
\concept{standardization functions}
