% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQC_Utils.R
\name{dataQC.LatitudeLongitudeCheck}
\alias{dataQC.LatitudeLongitudeCheck}
\title{find and standardize geographic coordinates in a dataframe}
\usage{
dataQC.LatitudeLongitudeCheck(dataset, latlon.colnames=list(c("lat_lon"),c("latitude"), c("longitude")))
}
\arguments{
\item{dataset}{dataframe. The dataset where the date column should be found}

\item{latlon.colnames}{a list of length 3 with character vectors. Three vectors of potential names for the columns with the latitude-longidude values. The first vector in the list are names if latitude and longitude were to be in the same column (e.g. the MIxS lat_lon format), the second and third are for when latitude and longitude, respectively, are in seperate columns. Example: list(c("lat_lon"), c("latitude"), c("longitude"))}
}
\value{
a list of length 2, with "$values" a vector of same length as the number of rows in the dataset argument, and "$warningmessages" a vector with potential warning messages as character strings.
}
\description{
looks in the columns of a dataset (dataframe) for a column with coordinates and transforms them a standardized decimal format (see details).
}
\details{
The date column is found based on a user-provided list of possible names to look for (latlon.colnames argument). First, a single column is searched where latitude and longitude are noted in a single field, if this returns no result, latitude and longitude are looked for in seperate fields. When found, the coordinates are transformed to decimals and returned as a single field, with values separated by a single space That is: (X Y), with X a numeric decimal latitude value and Y a numeric decimal longitude value.
}
\seealso{
Other quality control functions: 
\code{\link{dataQC.TaxonListFromData}()},
\code{\link{dataQC.TermsCheck}()},
\code{\link{dataQC.completeTaxaNamesFromRegistery}()},
\code{\link{dataQC.dateCheck}()},
\code{\link{dataQC.eventStructure}()},
\code{\link{dataQC.findNames}()},
\code{\link{dataQC.generate.footprintWKT}()},
\code{\link{dataQC.guess.env_package.from.data}()},
\code{\link{dataQC.taxaNames}()}
}
\author{
Maxime Sweetlove CC-0 2019
}
\concept{quality control functions}
