% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQC_Main_MIxS.R
\name{dataQC.MIxS}
\alias{dataQC.MIxS}
\title{format dataframes into a MIxS object}
\usage{
dataQC.MIxS(dataset = NA, ask.input=TRUE, add_to = NA, sample.names = NA)
}
\arguments{
\item{dataset}{data.frame. The raw dataset downloaded from INSDC to be cleaned up. Rows are samples, columns variables. Units can be listed in the first or second row, and will be automatically detected if the row names of this row includes the word "units". Different units per sample are not allowed.}

\item{ask.input}{logical. If TRUE, console input will be requested to the user when a problem occurs (process runs user-supervised). Default TRUE}

\item{add_to}{a MIxS.metadata object. An already present dataset with quality-controlled metadata. must be formatted as MIxS.metadata to ensure the correct input format of the data.}

\item{sample.names}{character. The column with sample names to use. Use row.names for rownames. If NA the function will try to find sample names itself. default NA}
}
\value{
a MIxS.metadata object that is compatible with the MIxS standard
}
\description{
takes a dataframe with contextual data and metadata from a sequencing dataset and performs a basis Quality Controll. (see details)
}
\details{
Any sequencing project typically has important additional data associated with it. This goes from laboratory protocols, sequencing platform settings or environmental measurements. Thisfunction was develloped to sort through these metadata (provided in a dataframe), and perform a basic quality controll, correcting the most common mistakes, like incorrectly formatting the geographic coordinates, formatting dates, typos or variants of variable names, etc. To do this, the function makes use of a build-in dictionary of (MIxS) terms and their synonyms (that is: spelling errors, writing differences, true synonyms,...). Note that it is possible some terms are not recognized. In that case contact the author of the package to update the dictionary in the upcomming version.
}
\seealso{
get.BioProject.metadata.INSDC, get.sample.attributes.INSDC

Other standardization functions: 
\code{\link{coordinate.to.decimal}()},
\code{\link{dataQC.DwC_general}()},
\code{\link{dataQC.DwC}()}
}
\author{
Maxime Sweetlove ccBY 4.0 2019
}
\concept{standardization functions}
