% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQC_Utils.R
\name{dataQC.TaxonListFromData}
\alias{dataQC.TaxonListFromData}
\title{extract the finest resolution taxonomy from a dataset}
\usage{
find.sampleTaxon(dataset)
}
\arguments{
\item{dataset}{a data.frame. The dataset with samples as rows, and taxonomy information in the columns. using the MIxS or DarwinCore taxonomy terms, the taxonomy information will be extracted}
}
\value{
a vector with the highest level taxonomic name found, with genus and species epithet separated by a space.
}
\description{
tries to find taxonomic names for samples (rows) in a dataset (data.frame),
}
\examples{
\donttest{
test_metadata <- data.frame(sample_name=paste("sample", 1:5, sep="_"),
                           genus=c("Aulacoseira", "Calothrix confervicola", "unknown species", "Micrasterias cf. denticulata", "Calothrix sp."),
                           row.names=paste("sample", 1:5, sep="_"))
dataQC.TaxonListFromData(test_metadata)
}
}
\seealso{
Other quality control functions: 
\code{\link{dataQC.LatitudeLongitudeCheck}()},
\code{\link{dataQC.TermsCheck}()},
\code{\link{dataQC.completeTaxaNamesFromRegistery}()},
\code{\link{dataQC.dateCheck}()},
\code{\link{dataQC.eventStructure}()},
\code{\link{dataQC.findNames}()},
\code{\link{dataQC.generate.footprintWKT}()},
\code{\link{dataQC.guess.env_package.from.data}()},
\code{\link{dataQC.taxaNames}()}
}
\author{
Maxime Sweetlove CC-0 2020
}
\concept{quality control functions}
