% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_SequenceData_INSDC.R
\name{download.sequences.INSDC}
\alias{download.sequences.INSDC}
\title{Download open sequence data to your computer.}
\usage{
download.sequences.INSDC(BioPrj = c(), destination.path = NA, apiKey=NA, unzip = FALSE, keep.metadata = TRUE, download.sequences = TRUE)
}
\arguments{
\item{BioPrj}{a list with character strings. A list of one or more BioProject numbers to be downloaded. Required argument.}

\item{destination.path}{a character string. The path to the directory where all the downloaded sequence data needs to go}

\item{apiKey}{a character string. Only required if download.sequences.INSDC(keep.metadata=TRUE). A personal API-key to the acces the NCBI databases, and required to use the Entrez Programming Utilities (E-utilities). An API-key (API stands for application programming interface) is a unique identifier used to authenticate a user. You can easily generate an API-key: see https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/}

\item{unzip}{boolean. If TRUE, the all *.fastq.gz files in the destination.path will unzipped. Default FALSE}

\item{keep.metadata}{boolean. If TRUE, the downloaded metadata can be saved to a file (Console), if FALSE it is discarded. Default TRUE}

\item{download.sequences}{boolean. If TRUE, the sequences will be downloaded to the destination.path. If FALSE, no sequences are downloaded. Default TRUE}
}
\value{
the sequence data are written to the destination.path, the metadata is returned as a data.frame to the console.
}
\description{
Downloads (high throughput) nucleotide sequence datasets that are deposited on the International Nucleotide Sequence Database Consortium (INSDC, e.g. SRA, ENA, GenBank,...). Any possible metadata and environmental data is also downloaded.
}
\details{
download.sequences.INSDC will write the sequence data (*.fastq.gz files) to a destiation path (e.g. a designated file), the metadata (if it should be kept) is written to the Console and should be caught in an R-varaiable that can be later written to a csv file by the user. Point of entry to INSDC is the SRA database from NCBI.
}
\examples{
\dontrun{
download.sequences.INSDC(BioPrj="PRJNA303951", destination.path=getwd(),
                         apiKey="YouPersonalAPIKey", unzip=FALSE,
                         keep.metadata = TRUE, download.sequences = TRUE)
}
}
\seealso{
Other downloading data functions: 
\code{\link{get.BioProject.metadata.INSDC}()},
\code{\link{get.sample.attributes.INSDC}()}
}
\author{
Maxime Sweetlove CC-0 2019
}
\concept{downloading data functions}
