% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Libraries.R
\docType{class}
\name{DwC.event-class}
\alias{DwC.event-class}
\title{an S4 class to document data formated in the DarwinCore event core format}
\description{
an S4 class to document data formated in the DarwinCore event core format
}
\section{Slots}{

\describe{
\item{\code{core}}{a data.frame. The DwC event core, formatted with samples/events as rows, and variables/DwCTerms as columns.}

\item{\code{occurrence}}{a data.frame. The occurrence extension, formatted with occurrences as rows and variables/DwCTerms as columns. Must contain "eventID", "occurrenceID" and "basisOfRecord"}

\item{\code{emof}}{a data.frame. The DwC extended Measurement or Fact (eMoF) extention. This can be used to store environmental data using the MIxS vocabulary.}

\item{\code{EML.url}}{a character. the url to the Ecological Metadata Language (EML) file from the IPT}

\item{\code{QC}}{boolean to indicate that the data been Quality controlled (TRUE) or not (FALSE)}
}}

