% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFormat_Utils.R
\name{eMoF.to.wideTable}
\alias{eMoF.to.wideTable}
\title{convert a DarwinCore extended Measurement or Fact (eMoF) file to a wide tabular format}
\usage{
eMoF.to.wideTable(dataset)
}
\arguments{
\item{dataset}{a dataframe of the eMoF file}
}
\value{
a list of length 3: "$data" the data in a wide formt, "$units" the units, and "$method" the methods
}
\description{
converts a DarwinCore extended Measurement Or Fact (eMOF) file, which is has a "long" file format into a wide tabular format
}
\details{
Long formated data if great for data arciving, but is difficult to use in day-to-day statistical analyses. 
  This function extracts the data from an eMoF file and puts it in a wide sample x variable table
}
\examples{
\donttest{
sampleNames <- c("sample_1", "sample_2")
test_emof <- data.frame(eventID = c(rep(c("sample_1", "sample_2"), 2)),
                        measurementType = factor(c(rep("var1", 2), rep("var2", 2))), 
                        measurementValue = c(1:4),
                        measurementUnit = c(rep("unit1", 2), rep("unit2", 2)))
eMoF.to.wideTable(test_emof)
}
}
\seealso{
Other formating functions: 
\code{\link{combine.data.frame}()},
\code{\link{combine.data}()},
\code{\link{wideTab.to.hierarchicalTab}()},
\code{\link{wideTable.to.eMoF}()}
}
\author{
Maxime Sweetlove
}
\concept{formating functions}
