% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsMetaData.R
\docType{data}
\name{ENA_geoloc}
\alias{ENA_geoloc}
\title{geographic locations names accepted by ENA-EMBL}
\description{
geographic locations names accepted by ENA-EMBL
}
\references{
\url{https://www.ebi.ac.uk/ena/browser/home}
}
\author{
Maxime Sweetlove CC-0 2021
}
\keyword{ENA}
\keyword{INSDC,}
\keyword{data,}
