% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_sequenceData_ENA.R
\name{FileNames.to.Table}
\alias{FileNames.to.Table}
\title{collect the names of sequence files in a folder}
\usage{
FileNames.to.Table(file.dir, paired=TRUE, seq.file.extension=".fastq.gz", 
  pairedEnd.extension=c("_1", "_2"))
}
\arguments{
\item{file.dir}{a character string. The path to the directory where the sequence files are stored}

\item{paired}{boolean. wether or not the sequence files are paired-end (forward _1, reverse_2) or single-end}

\item{seq.file.extension}{a character string. The file-extension of the sequence files}

\item{pairedEnd.extension}{a character vector of length 2. If the data is paired-end data, specify the forward (first element of the vector) and reverse (second) extension tags here. Default is c("_1", "_2")}
}
\value{
a data frame with a column "OldName" that list the sequence file names, an empty column "NewName" to be filled in by the user and a column "FileName" with the full file name.
}
\description{
makes a table with the all the names of sequence files in a folder. see details.
}
\details{
fastq files from sequencing facilities often come with long and complex file names that were automatically generated by the sequencer machine and no longer resemble the original name of the sample. 
  This function is part of tools that help to get file names and easily convert them back into the original file names. It makes a table with the all the names of sequence files in a folder that can be saved as a CSV file.
  Using a text editor or excell, the user can then fill in the column with new name (that is, the desired name), and then use the renameSequenceFiles() function to rename the files.
}
\examples{
\donttest{
FileNames.to.Table(file.dir="path/to/the/sequenceFilesFolder", paired=TRUE, 
                   seq.file.extension=".fastq.gz",
                   pairedEnd.extension=c("_1", "_2"))
}
}
\seealso{
Other data archiving functions: 
\code{\link{commonTax.to.NCBI.TaxID}()},
\code{\link{get.ENAName}()},
\code{\link{prep.metadata.ENA}()},
\code{\link{renameSequenceFiles}()},
\code{\link{sync.metadata.sequenceFiles}()}
}
\author{
Maxime Sweetlove
}
\concept{data archiving functions}
