% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Utils.R
\name{find.dataset}
\alias{find.dataset}
\title{find a column in a dataframe}
\usage{
find.dataset(dataset, TermsList = c())
}
\arguments{
\item{dataset}{data.frame. A dataframe where to find a column based on the columnames}

\item{TermsList}{a character vector with the columnnames to look for. They will be handled in the order given, and the search will stop at the first match}
}
\value{
a character vactor with the content of the matching column (or the first match), or an empty vector
}
\description{
find a column in a dataframe where it is a priori unkown wether the column is present or not.
}
\details{
An internal function to find information in incomming data of which the format and content is unknown
}
\examples{
\donttest{
find.dataset(dataset=data.frame(var1=1:3, var2=1:3), "var2")
}
}
\author{
Maxime Sweetlove
}
