% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Utils.R
\name{get.boundingBox}
\alias{get.boundingBox}
\title{get a rectagular bounding box from a set of coordinates}
\usage{
get.boundingBox(latitudes, longitudes)
}
\arguments{
\item{latitudes}{numeric vector. one ore more decimal latitude values}

\item{longitudes}{numeric vector. one ore more decimal longitude values}
}
\value{
a character string with the boundingbox printed to the console
}
\description{
finds the (rectangular) bounding box given decimal latitudes and longitudes
}
\details{
Documenting data with metadata on an Integrated Publishing Toolkit (IPT) typically requires a bounding box to simplify the goegraphic content of the dataset. This function is made to quickly calculate this.
}
\author{
Maxime Sweetlove CC-0 2019
}
