% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Utils.R
\name{get.ENAName}
\alias{get.ENAName}
\title{convert a MIxS term into it's ENA variant}
\usage{
get.ENAName(variable)
}
\arguments{
\item{variable}{character a MIxS term.}
}
\value{
a character
}
\description{
get the ENA variant a MIxS term
}
\examples{
\donttest{
get.ENAName("target_gene")
}
}
\seealso{
Other data archiving functions: 
\code{\link{FileNames.to.Table}()},
\code{\link{prep.metadata.ENA}()},
\code{\link{renameSequenceFiles}()},
\code{\link{sync.metadata.sequenceFiles}()}
}
\author{
Maxime Sweetlove
}
\concept{data archiving functions}
