% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Utils.R
\name{get.insertSize}
\alias{get.insertSize}
\title{Get the sequence length of the first sequence in a fastq or fastq.gz file}
\usage{
get.insertSize(file_path = NA)
}
\arguments{
\item{file_path}{the full file path to the file}
}
\value{
a numeric value
}
\description{
get the insert size (i.e. number of bases) of a fastq or fastq.gz file
}
\details{
submitting sequence to a database of the Nucleotide Sequence Database Consortium typically requires to give the insert size of the sequences. This function gets the length of the first sequence in a file, assuming all sequences are of the same length.
}
\examples{
\dontrun{
get.insertSize("/user/path/to/sequenceFileFolder")
}
}
\author{
Maxime Sweetlove CC-0 2020
}
