% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_SequenceData_INSDC.R
\name{get.sample.attributes.INSDC}
\alias{get.sample.attributes.INSDC}
\title{Downloads all sequence sample attributes}
\usage{
get.sample.attributes.INSDC(sampleID, apiKey, BioPrjct)
}
\arguments{
\item{sampleID}{a list. a list of one ore more SRA sample IDs (that is "Run" numbers). This argument can be left blank if input is provided for the BioPrjct argument (see further)}

\item{apiKey}{a character string. A personal API-key to the acces the NCBI databases, and required to use the Entrez Programming Utilities (E-utilities). An API-key (API stands for application programming interface) is a unique identifier used to authenticate a user. A personal API-key can easily be generated at requested at https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/}

\item{BioPrjct}{a character string or a vector with character strings. Providing the associated BioProject numbers helps to give more understandable error messages. Alternativey, if the sampleID argument is empty, all sample IDs from the given BioProject numbers will be given to the sampleID argument.}
}
\value{
a dataframe with the data found
}
\description{
Downloads all sample attributes (that is, additional environmental or other associated data) from INSDC. Note: requires a user-specified API-key to acces the INSDC databases. see https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/ to generate an API-key
}
\details{
Each sequence data sample ("Run") typically has additional measurements or metadata associated with it. However, these are difficult to find, and cannot be downloaded together with the nucleotide sequence data. The get.sample.attributes.INSDC function will fetch the all the metadata of the samples given to the sampleID (or BioPrjct) argument from the INSDC. To do this, get.sample.attributes.INSDC will use the Entrez Programming Utilities (E-utilities) from NCBI, and will access the databases with the API-key provided by the user. see https://ncbiinsights.ncbi.nlm.nih.gov/2017/11/02/new-api-keys-for-the-e-utilities/ for more info on getting an API-key.
}
\seealso{
Other downloading data functions: 
\code{\link{download.sequences.INSDC}()},
\code{\link{get.BioProject.metadata.INSDC}()}
}
\author{
Maxime Sweetlove CC-0 2019
}
\concept{downloading data functions}
