% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Libraries.R
\docType{class}
\name{MIxS.metadata-class}
\alias{MIxS.metadata-class}
\title{an S4 class to document data formated in the MIxS standard}
\description{
an S4 class to document data formated in the MIxS standard
}
\section{Slots}{

\describe{
\item{\code{data}}{a data.frame with the meta- and/or environmental data, formatted with samples/events as rows, and variables/MIxS terms as columns.}

\item{\code{section}}{a character, a vector with the same length and order as number of columns of the dataframe in the data slot, containing a MIxS section/cathegory for each term}

\item{\code{units}}{a character. a vector with the same length and order as number of columns of the dataframe in the data slot, containing a unit for each variable (use "alphanumeric" if there is no unit)}

\item{\code{type}}{a character. Either "versatile", meaning the object is a loose following of the MIxS standard, or "strict.MIxS", meaning the object strictly follows all the MIxS rules.}

\item{\code{env_package}}{a character. The MIxS environmental package that is appropriate for this dataset. Can also be not_specified or multiple_packages.}

\item{\code{QC}}{boolean to indicate that the data been Quality controlled (TRUE) or not (FALSE)}
}}

