% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Utils.R
\name{multi.warnings}
\alias{multi.warnings}
\title{collect warning messages troughout a running function}
\usage{
multi.warnings(message_text, warningmessages)
}
\arguments{
\item{message_text}{a character string. The error message to be added}

\item{warningmessages}{a vector with one ore more character strings. The previous error messages.}
}
\value{
a character string with unique warning messages
}
\description{
collect warning messages troughout a running function
}
\author{
Maxime Sweetlove CC-0 2019
}
