% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Libraries.R
\name{show,DwC.event-method}
\alias{show,DwC.event-method}
\title{the show method for DwC.event}
\usage{
\S4method{show}{DwC.event}(object)
}
\description{
the show method for DwC.event
}
\author{
Maxime Sweetlove ccBY 4.0 2020
}
