% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Libraries.R
\name{show,DwC.occurrence-method}
\alias{show,DwC.occurrence-method}
\title{the show method for DwC.occurrence}
\usage{
\S4method{show}{DwC.occurrence}(object)
}
\arguments{
\item{object}{a DwC.occurrence object}
}
\description{
the show method for DwC.occurrence
}
\author{
Maxime Sweetlove CC-0 2020
}
