% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Libraries.R
\name{show,MIxS.metadata-method}
\alias{show,MIxS.metadata-method}
\title{the show method for MIxS.metadata}
\usage{
\S4method{show}{MIxS.metadata}(object)
}
\description{
the show method for MIxS.metadata
}
\author{
Maxime Sweetlove CC-0 2020
}
