% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_sequenceData_ENA.R
\name{sync.metadata.sequenceFiles}
\alias{sync.metadata.sequenceFiles}
\title{Check if all samples in a dataframe have sequence data}
\usage{
sync.metadata.sequenceFiles(Names, file.dir=NULL, 
  paired=TRUE, seq.file.extension=".fastq.gz", 
  pairedEnd.extension=c("_1", "_2"))
}
\arguments{
\item{Names}{a vector or a MIxS.metadata class object. A character vector with the sample names (without a file extension) or the MIxS.metadata object of which the sample names (original_name) should be compared to the sequence file names}

\item{file.dir}{a character string. The path to the directory where the sequence files are stored}

\item{paired}{boolean. wether or not the sequence files are paired-end (forward _1, reverse_2) or single-end}

\item{seq.file.extension}{a character string. The file-extension of the sequence files}

\item{pairedEnd.extension}{a character vector of length 2. If the data is paired-end data, specify the forward (first element of te vector) and reverse (second) extension tags here. Default is c("_1", "_2")}
}
\value{
a list of two. Two lists with the non-matching file names. redundant_metadata shows the names found in the metadata but not among the sequences, redundant_seqFiles shows sequence file names that were not listed in the metadata.
}
\description{
checks if all records in the metadata file match with filenames in a given directory
}
\details{
Nucleotide sequence datasets typically contain a large number of samples, too much to check manually. This function will check if all the names that are expected in the sequence data files are actually found. Reasons for mismatches may include typos, accedentally deleting files or records, errors in naming or copying files,...
}
\examples{
\donttest{
sync.metadata.sequenceFiles(Names="seq_sample1", file.dir="user/path/to/sequenceFilesFolder",
                            paired=TRUE, seq.file.extension=".fastq.gz",
                            pairedEnd.extension=c("_1", "_2"))
}
}
\seealso{
Other data archiving functions: 
\code{\link{FileNames.to.Table}()},
\code{\link{commonTax.to.NCBI.TaxID}()},
\code{\link{get.ENAName}()},
\code{\link{prep.metadata.ENA}()},
\code{\link{renameSequenceFiles}()}
}
\author{
Maxime Sweetlove
}
\concept{data archiving functions}
