% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsMetaData.R
\docType{data}
\name{TaxIDLib}
\alias{TaxIDLib}
\title{non-exaustive library with some common INSDC taxon IDs}
\description{
non-exaustive library with some common INSDC taxon IDs
}
\references{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Browser/wwwtax.cgi}
}
\author{
Maxime Sweetlove CC-0 2021
}
\keyword{INSDC,}
\keyword{NCBI,}
\keyword{data,}
\keyword{taxonomy}
