% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsMetaData.R
\name{term.definition}
\alias{term.definition}
\title{Find the MIxS or DarwinCore standard term and definition of a variable}
\usage{
term.definition(term)
}
\arguments{
\item{term}{a character string. The variable to look for among the MIxS and DarwinCore vocabularies}
}
\value{
chracater string printed to the console. The best matching terms and their the definitions.
}
\description{
retrieve the MIxS or DarwinCore standard term and definition of a variable.
}
\details{
Standerdizing microbial sequence data, metadata and environmental data can be quite difficult given the plethora of standard terms already in existance. This function returns a definition of any term that is used on the POLAAAR portal at biodiversity.aq.
}
\author{
Maxime Sweetlove CC-0 2019
}
