% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsMetaData.R
\docType{data}
\name{TermsLib}
\alias{TermsLib}
\title{central library with mapped terms of DwC, MIxS and miscellaneous missing terms}
\description{
central library with mapped terms of DwC, MIxS and miscellaneous missing terms
}
\references{
\url{https://gensc.org/mixs/} \url{https://dwc.tdwg.org}
}
\author{
Maxime Sweetlove CC-0 2021
}
\keyword{DarwinCore}
\keyword{MIxS,}
\keyword{data,}
