% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsMetaData.R
\docType{data}
\name{TermsSyn}
\alias{TermsSyn}
\title{library with synonyms for standard terms}
\description{
library with synonyms for standard terms
}
\references{
\url{https://gensc.org/mixs/} \url{https://dwc.tdwg.org}
}
\author{
Maxime Sweetlove CC-0 2021
}
\keyword{DarwinCore,}
\keyword{ENA,}
\keyword{INSDC,}
\keyword{MIxS,}
\keyword{NCBI,}
\keyword{POLA3R}
\keyword{data,}
