% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmicsMetaData.R
\docType{data}
\name{TermsSyn_DwC}
\alias{TermsSyn_DwC}
\title{library with synonyms for DwC terms}
\description{
library with synonyms for DwC terms
}
\references{
\url{https://dwc.tdwg.org}
}
\author{
Maxime Sweetlove CC-0 2021
}
\keyword{DarwinCore}
\keyword{data,}
