% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFormat_Utils.R
\name{wideTab.to.hierarchicalTab}
\alias{wideTab.to.hierarchicalTab}
\title{convert a wide table to hierarchical table}
\usage{
wideTab.to.hierarchicalTab(dataset, col_hierarchy)
}
\arguments{
\item{dataset}{a data.frame. The wide table to be transformed, with columns as hierarchical variables. T}

\item{col_hierarchy}{a data.frame with two columns named "child" and "parent". The hierarchical relations between the columns formatted in a child-parent table listing the column names. Use "root" or NA for columns with no parent.}
}
\value{
a data.frame with three columns: one for the value (cell content in the origical data.frame), it's rank (the original column name), and parent (cell content of the column that is one up in the hierarchy)
}
\description{
turns a regular wide table into a hierarchical recursive table.
}
\details{
Columns in the input data with different levels of a hierarchy are poolled into one column with values and a second column indicating the parent of the value.
}
\seealso{
Other formating functions: 
\code{\link{combine.data.frame}()},
\code{\link{combine.data}()},
\code{\link{eMoF.to.wideTable}()},
\code{\link{wideTable.to.eMoF}()}
}
\author{
Maxime Sweetlove CC-0 2019
}
\concept{formating functions}
