% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Libraries.R
\name{write.MIxS}
\alias{write.MIxS}
\title{Data Output: write a MIxS.metadata class object to a CSV file}
\usage{
write.MIxS(x, file="")
}
\arguments{
\item{x}{the MIxS.metadata class object to be written.}

\item{file}{either a character string naming a file or a connection open for writing. "" indicates output to the console.}
}
\description{
write a MIxS.metadata class object to a CSV file
}
\author{
Maxime Sweetlove CC-0 2020
}
