% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFormat_Utils.R
\name{combine.data.frame}
\alias{combine.data.frame}
\title{merge dataframes}
\usage{
combine.data.frame(df1, df2, fill=NA, merge.cols=TRUE, original_rowName.col=TRUE, merge.rows="df1")
}
\arguments{
\item{df1}{a dataframe}

\item{df2}{a dataframe}

\item{fill}{character or NA. A value to put into the cells that have no data. Default NA}

\item{merge.cols}{logical. Merge colomns with common name. Default TRUE}

\item{original_rowName.col}{logical. if TRUE, add a column with the original rownames. Default TRUE}

\item{merge.rows}{character. If the data.frames df1 and df2 have common rows, the values for these rows of this dataset must be kept in favor of the other. If NA, both rows will be kept. Either "df1", "df2" or NA. Default "df1"}
}
\value{
a data.frame
}
\description{
combine.data.frame merges two dataframes, completing the rows and columns that are not shared by the dataframes.
}
\details{
Columns with matching names can or cannot be merged, rows are automatically bound (a wrapper of rbind), not merged. Any missing data as a result of the non-matchng columns will be filled by the fill argument.
}
\examples{
\donttest{
df01 <- data.frame(var1=1:4, var2=2:5, row.names=paste("r", 1:4, sep=""))
df02 <- data.frame(var1=5:8, var3=1:4, row.names=paste("t", 1:4, sep=""))
combine.data.frame(df1=df01, df2=df02, fill=NA, merge.cols=TRUE, 
                   original_rowName.col=TRUE, merge.rows="df1")
}
}
\seealso{
Other formating functions: 
\code{\link{combine.data}()},
\code{\link{eMoF.to.wideTable}()},
\code{\link{wideTab.to.hierarchicalTab}()},
\code{\link{wideTable.to.eMoF}()}
}
\author{
Maxime Sweetlove CC-0 2019
}
\concept{formating functions}
