% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQC_Utils.R
\name{dataQC.findNames}
\alias{dataQC.findNames}
\title{find the samplenames in a dataset}
\usage{
dataQC.findNames(dataset, ask.input=TRUE, sample.names=NA)
}
\arguments{
\item{dataset}{data.frame. The data.frame where to look for the sample names}

\item{ask.input}{logical. If TRUE, console input will be requested to the user when a problem occurs. Default TRUE}

\item{sample.names}{character. The column with sample names to use. Use row.names for rownames. If NA the function will try to find sample names itself. default NA}
}
\value{
a list of length 3 with: "$Names" a named vector with the most likely sample names, "$Names.column" the column name where the sample names were found, "$warningmessages" a vector with warning messages
}
\description{
find the sample names in a given (meta-)dataset where at least an attempt has been made to standardize the data following MIxS or DarwinCore
}
\details{
It is often not clear where the sample names are in a dataset. This function makes an educated guess, based on rownames or tags that are often used to indicate sample names. If ask.input, then the process happens user-supervised.
}
\examples{
\donttest{
test_metadata <- data.frame(sample_name=paste("sample", 1:5, sep="_"),
                           collection_date=c("2020-09-23", "2020", "16 Jan. 2020", "November 1998", "12/01/1999"),
                           latitude=c(23, 45, -56.44, "47.5", "-88° 4\' 5\""),
                           longitude=c(24, -57, -107.55, "33.5", "-130° 26\' 9\""),
                           row.names=paste("sample", 1:5, sep="_"))
dataQC.findNames(dataset=test_metadata, ask.input=TRUE, sample.names="sample_name")
}
}
\seealso{
Other quality control functions: 
\code{\link{dataQC.LatitudeLongitudeCheck}()},
\code{\link{dataQC.TaxonListFromData}()},
\code{\link{dataQC.TermsCheck}()},
\code{\link{dataQC.completeTaxaNamesFromRegistery}()},
\code{\link{dataQC.dateCheck}()},
\code{\link{dataQC.eventStructure}()},
\code{\link{dataQC.generate.footprintWKT}()},
\code{\link{dataQC.guess.env_package.from.data}()},
\code{\link{dataQC.taxaNames}()}
}
\author{
Maxime Sweetlove CC-0 2020
}
\concept{quality control functions}
