% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQC_Utils.R
\name{dataQC.guess.env_package.from.data}
\alias{dataQC.guess.env_package.from.data}
\title{make an educated guess to the MIxS environmental package of a dataset}
\usage{
dataQC.guess.env_package.from.data(dataset, 
  pckge.colnames=c("env_package", "ScientificName"))
}
\arguments{
\item{dataset}{dataframe. The dataset for which the MIxS environmental package should be found or guessed}

\item{pckge.colnames}{a character vector. A vector with the potential names for the column where the environmental package can be found. Place the terms in order of decreasing likeliness.}
}
\value{
a list of length 2, with "$values" a vector of same length as the number of rows in the dataset argument, and "$warningmessages" a vector with potential warning messages as character strings.
}
\description{
looks in the columns of a dataset for clues to what the most appropriate MIxS environmental could be for each sample.
}
\details{
The "Minimum Information on any (x) Sequence" (MIxS) standard requires an appropriate environmental package with MIxS terms to be selected to document the data. 
  Some data and nucleotide archives enforce their users to select such a package. This function is made to automatically either find the package in a dataset, of guess it based on the data that is present.
}
\examples{
\donttest{
test_metadata <- data.frame(sample_name=paste("sample", 1:5, sep="_"),
                           collection_date=c("2020-09-23", 
                                             "2020", 
                                             "16 Jan. 2020", 
                                             "November 1998", 
                                             "12/01/1999"),
                           env_package=rep("water", 5),
                           row.names=paste("sample", 1:5, sep="_"))
dataQC.guess.env_package.from.data(test_metadata, 
                                   pckge.colnames=c("env_package"))
}
}
\seealso{
Other quality control functions: 
\code{\link{dataQC.LatitudeLongitudeCheck}()},
\code{\link{dataQC.TaxonListFromData}()},
\code{\link{dataQC.TermsCheck}()},
\code{\link{dataQC.completeTaxaNamesFromRegistery}()},
\code{\link{dataQC.dateCheck}()},
\code{\link{dataQC.eventStructure}()},
\code{\link{dataQC.findNames}()},
\code{\link{dataQC.generate.footprintWKT}()},
\code{\link{dataQC.taxaNames}()}
}
\author{
Maxime Sweetlove CC-0 2019
}
\concept{quality control functions}
