% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes_Libraries.R
\name{write.MIxS}
\alias{write.MIxS}
\title{Data Output: write a MIxS.metadata class object to a CSV file}
\usage{
write.MIxS(x, file="")
}
\arguments{
\item{x}{the MIxS.metadata class object to be written.}

\item{file}{either a character string naming a file or a connection open for writing. "" indicates output to the console.}
}
\description{
write a MIxS.metadata class object to a CSV file
}
\examples{
\dontrun{
sampleNames <- c("sample_1", "sample_2")
test_MIxS <- new("MIxS.metadata",
                 data = data.frame(var1=c(1,2), var2=c(3,4), eventID=sampleNames, 
                                   row.names=sampleNames),
                 section = c(var1="section1", var2="section1", eventID="miscellaneous"),
                 units      = c(var1="unit1", var2="unit2", eventID="alphanumeric"),
                 env_package = "water",
                 type = "versatile",
                 QC = TRUE)
write.MIxS(x=test_MIxS, file="test.csv")
}
}
\author{
Maxime Sweetlove CC-0 2020
}
